character class kinda done

+++ Sprint 1 (V1.1)/Block.java
@@ -34,7 +34,7 @@ public class Block {
       return this.y;
    }
    
-   public double getDurability() {
+   public int getDurability() {
       return this.durability;
    }
